requirements-gpu.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
onnxruntime-gpu==1.18.1
python-multipart==0.0.9
python-json-logger==2.0.7
google-cloud-storage==2.12.0

